fix: dont rerender the page when clicking an active route when on mobile
